requirements.txt
apache-airflow-providers-http==4.9.1  
apache-airflow-providers-snowflake==5.3.1 
snowflake-connector-python==3.7.1 
snowflake-sqlalchemy==1.5.1
pandas==2.0.3  
requests==2.31.0 
boto3==1.34.39
botocore==1.34.39  
